Config Generation: Extract post-processing code Put all of that code into a separate dir
